+++ src/Controller/ArticleReferenceAdminController.php
@@ -10,6 +10,7 @@ use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\HttpFoundation\HeaderUtils;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\StreamedResponse;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Validator\Constraints\File;
@@ -56,14 +57,20 @@ class ArticleReferenceAdminController extends BaseController
             ]
         );
         if ($violations->count() > 0) {
+
             /** @var ConstraintViolation $violation */
+            /* Old way to redirect and show errors, now we treat our method like an API endpoint
             //violations is an object but implements ArrayAccess
-            $violation = $violations[0];
-            //fast way to push one-time error message
-            $this->addFlash('error', $violation->getMessage());
-            return $this->redirectToRoute('admin_article_edit', [
-                'id' => $article->getId(),
-            ]);
+                $violation = $violations[0];
+                //fast way to push one-time error message
+                $this->addFlash('error', $violation->getMessage());
+                return $this->redirectToRoute('admin_article_edit', [
+                    'id' => $article->getId(),
+                ]);
+             */
+
+            // we return object by json
+            return $this->json($violations, 400);
         }
         // End of Validation
 
@@ -75,9 +82,38 @@ class ArticleReferenceAdminController extends BaseController
         $articleReference->setMimeType($uploadedFile->getMimeType() ?? 'application/octet-stream');
         $entityManager->persist($articleReference);
         $entityManager->flush();
-        return $this->redirectToRoute('admin_article_edit', [
-            'id' => $article->getId(),
-        ]);
+
+        /* Old way to redirect and show errors, now we treat our method like an API endpoint
+            return $this->redirectToRoute('admin_article_edit', [
+                'id' => $article->getId(),
+            ]);
+        */
+
+        return $this->json(
+            $articleReference,
+            201,
+            [],
+            [ // we use group because we have an article in articleReference and article contain articleReference so we need to avoid endless loop while serialization
+                'groups' => ['main']
+            ]
+        );
+}
+
+    /**
+     * API endpoint to get current list of references
+     * @Route("/admin/article/{id}/references", methods="GET", name="admin_article_list_references")
+     * @IsGranted("MANAGE", subject="article")
+     */
+    public function getArticleReferences(Article $article)
+    {
+        return $this->json(
+            $article->getArticleReferences(),
+            200,
+            [],
+            [
+                'groups' => ['main']
+            ]
+        );
     }
 
     /**
@@ -118,5 +154,25 @@ class ArticleReferenceAdminController extends BaseController
         $response->headers->set('Content-Disposition', $disposition);
 
         return $response;
+    }
+
+    /**
+     * Handle deleting elements from references list by API Endpoint
+     * @Route("/admin/article/references/{id}", name="admin_article_delete_reference", methods={"DELETE"})
+     */
+    public function deleteArticleReference(ArticleReference $reference, UploaderHelper $uploaderHelper, EntityManagerInterface $entityManager)
+    {
+        // from the reference we get article to check if we has an access to this article
+        $article = $reference->getArticle();
+        $this->denyAccessUnlessGranted('MANAGE', $article);
+
+        // we remvoe reference
+        $entityManager->remove($reference);
+        $entityManager->flush();
+
+        //remove file from reference
+        $uploaderHelper->deleteFile($reference->getFilePath(), false);
+
+        return new Response(null, 204);
     }
 }